move tests near implementation
