+++ packages/detectors/src/detectors/browser/query-string.test.ts
@@ -1,7 +1,7 @@
 import { suite } from 'uvu'
 import * as assert from 'uvu/assert'
-import type { Locale } from '../../runtime/src/core.mjs'
-import { initQueryStringDetector } from '../../src/detectors/browser/query-string.mjs'
+import type { Locale } from '../../../../runtime/src/core.mjs'
+import { initQueryStringDetector } from './query-string.mjs'
 
 const test = suite('detector:query-string')
