+++ packages/detectors/src/detectors/server/request-hostname.test.ts
@@ -1,7 +1,7 @@
 import { suite } from 'uvu'
 import * as assert from 'uvu/assert'
-import type { Locale } from '../../runtime/src/core.mjs'
-import { initRequestHostnameDetector } from '../../src/detectors/server/request-hostname.mjs'
+import type { Locale } from '../../../../runtime/src/core.mjs'
+import { initRequestHostnameDetector } from './request-hostname.mjs'
 
 const test = suite('detector:request-hostname')
