+++ packages/detectors/tsconfig.json
@@ -3,5 +3,5 @@
 	"compilerOptions": {
 		"outDir": "../../detectors"
 	},
-	"include": ["src/**/*.mts"]
+	"include": ["src/**/*.mts", "src/**/*.ts"]
 }
